+++ SnipeSharp/EndPoint/Models/Component.cs
@@ -10,7 +10,7 @@ namespace SnipeSharp.EndPoint.Models
     public class Component : CommonEndPointModel
     {
         [Field("id")]
-        public override long Id { get; set; }
+        public override int Id { get; set; }
 
         [Field("name")]
         public override string Name { get; set; }
